Move tc/core/polyhedral/functional.h tc/core/functional.h
The functional abstraction is more generally useful than just in polyhedral
